pyproject.toml
[tool.poetry]
name = "maple-guild-manage-api-server"
version = "0.1.0"
description = ""
authors = ["changsu lee <leecs0503@naver.com>"]
readme = "README.md"
packages = [{include = "maple_guild_manage_api_server"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
pymysql = "^1.0.3"
fastapi = "^0.103.0"
numpy = "^1.26.0"
pillow = "^10.0.1"
uvicorn = "^0.23.2"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pytest = "^7.4.0"
ipykernel = "^6.25.2"
pyright = "^1.1.327"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
